+++ src/clj/lambdacd/core.clj
@@ -1,7 +1,7 @@
 (ns lambdacd.core
   (:use compojure.core)
   (:require [lambdacd.internal.default-pipeline-state :as default-pipeline-state]
-            [lambdacd.internal.execution :as execution]
+            [lambdacd.execution.core :as execution]
             [lambdacd.event-bus :as event-bus]
             [lambdacd.internal.running-builds-tracking :as running-builds-tracking]
             [lambdacd.state.internal.pipeline-state-updater :as pipeline-state-updater]
@@ -9,7 +9,8 @@
             [clojure.tools.logging :as log]
             [lambdacd.runners :as runners]
             [lambdacd.execution.internal.execute-steps :as execute-steps]
-            [lambdacd.execution.internal.kill :as kill]))
+            [lambdacd.execution.internal.kill :as kill]
+            [lambdacd.execution.core :as execution-core]))
 
 (defn- add-shutdown-sequence! [ctx]
   (doto (Runtime/getRuntime)
@@ -48,22 +49,23 @@
      {:context                context
       :pipeline-def           pipeline-def})))
 
-(defn retrigger [pipeline context build-number step-id-to-retrigger]
-  "DEPRECATED, use lambdacd.execution instead"
-  (execution/retrigger-async pipeline context build-number step-id-to-retrigger))
+(defn retrigger
+  "DEPRECATED, use lambdacd.execution.core/retrigger instead"
+  [pipeline context build-number step-id-to-retrigger]
+  (execution-core/retrigger-pipeline-async pipeline context build-number step-id-to-retrigger))
 
 (defn kill-step [ctx build-number step-id]
-  "DEPRECATED, use lambdacd.execution instead"
-  (kill/kill-step ctx build-number step-id))
+  "DEPRECATED, use lambdacd.execution.core/kill-step instead"
+  (execution-core/kill-step ctx build-number step-id))
 
 (defn execute-steps [steps args ctx & opts]
-  "DEPRECATED, use lambdacd.execution instead"
-  (apply execute-steps/execute-steps steps args ctx opts))
+  "DEPRECATED, use lambdacd.execution.core instead"
+  (apply execution-core/execute-steps steps args ctx opts))
 
 (defn execute-step
   ([args ctx step]
-   "DEPRECATED, use lambdacd.execution instead"
-   (execute-step/execute-step args [ctx step]))
+   "DEPRECATED, use lambdacd.execution.core instead"
+   (execution-core/execute-step args [ctx step]))
   ([args [ctx step]]
-   "DEPRECATED, use lambdacd.execution instead"
-   (execute-step/execute-step args [ctx step])))
+   "DEPRECATED, use lambdacd.execution.core instead"
+   (execution-core/execute-step args [ctx step])))