+++ src/clj/lambdacd/execution.clj
@@ -1,26 +1,31 @@
 (ns lambdacd.execution
-  "Public interface to interact with the execution-engine"
-  (:require [lambdacd.internal.execution :as internal]
-            [lambdacd.execution.internal.execute-step :as execute-step]
-            [lambdacd.execution.internal.execute-steps :as execute-steps]
-            [lambdacd.execution.internal.kill :as kill]))
+  "DEPRECATED, use lambdacd.execution.core instead"
+  (:require [lambdacd.execution.core :as execution-core]))
 
-; TODO: we can probably deprecate this in favor of a package like lambdacd.execution.core for more consistency (like lambdacd.state.core)
-
-(defn retrigger [pipeline context build-number step-id-to-retrigger]
-  (internal/retrigger-async pipeline context build-number step-id-to-retrigger))
+(defn retrigger
+  "DEPRECATED, use lambdacd.execution.core/retrigger-pipeline-async instead"
+  [pipeline context build-number step-id-to-retrigger]
+  (execution-core/retrigger-pipeline-async pipeline context build-number step-id-to-retrigger))
 
 (defn kill-step [ctx build-number step-id]
-  (kill/kill-step ctx build-number step-id))
+  "DEPRECATED, use lambdacd.execution.core/kill-step instead"
+  (execution-core/kill-step ctx build-number step-id))
 
 (defn execute-steps [steps args ctx & opts]
-  (apply execute-steps/execute-steps steps args ctx opts))
+  "DEPRECATED, use lambdacd.execution.core/execute-steps instead"
+  (apply execution-core/execute-steps steps args ctx opts))
 
 (defn execute-step
   ([args ctx step]
-   (execute-step/execute-step args [ctx step]))
+   "DEPRECATED, use lambdacd.execution.core/execute-steps instead"
+   (execution-core/execute-step args [ctx step]))
   ([args [ctx step]]
-   (execute-step/execute-step args [ctx step])))
+   "DEPRECATED, use lambdacd.execution.core/execute-steps instead"
+   (execution-core/execute-step args [ctx step])))
 
-(defn run [pipeline context]
-  (internal/run pipeline context))
+(defn run
+  "DEPRECATED, use lambdacd.execution.core/run-pipeline instead"
+  [pipeline context]
+  (execution-core/run-pipeline pipeline context))
+
+