+++ src/clj/lambdacd/execution/internal/execute_steps.clj
@@ -5,18 +5,9 @@
             [lambdacd.util.internal.sugar :refer [not-nil?]]
             [lambdacd.execution.internal.serial-step-result-producer :refer [serial-step-result-producer]]
             [lambdacd.step-id :as step-id]
-            [lambdacd.state.core :as state]
-            [lambdacd.execution.internal.util :as execution-util]))
-
-(defn- sequential-retrigger-predicate [ctx step]
-  (let [cur-step-id         (:step-id ctx)
-        retriggered-step-id (:retriggered-step-id ctx)]
-    (cond
-      (or
-        (step-id/parent-of? cur-step-id retriggered-step-id)
-        (= cur-step-id retriggered-step-id)) :rerun
-      (step-id/later-than? cur-step-id retriggered-step-id) :run
-      :else :mock)))
+            [lambdacd.execution.internal.util :as execution-util]
+            [lambdacd.execution.internal.kill :as kill]
+            [lambdacd.execution.internal.retrigger :as retrigger]))
 
 (defn unify-only-status [unify-status-fn]
   (fn [step-results]
@@ -62,38 +53,6 @@
             (recur new-results))
           (async/close! dropping-output-ch))))))
 
-
-(defn- publish-child-step-results!! [ctx retriggered-build-number original-build-result]
-  (->> original-build-result
-       (filter #(step-id/parent-of? (:step-id ctx) (first %)))
-       (map #(execution-util/send-step-result!! (assoc ctx :step-id (first %)) (assoc (second %) :retrigger-mock-for-build-number retriggered-build-number)))
-       (doall)))
-
-(defn retrigger-mock-step [retriggered-build-number]
-  (fn [args ctx]
-    (let [original-build-result (state/get-step-results ctx retriggered-build-number)
-          original-step-result  (get original-build-result (:step-id ctx))]
-      (publish-child-step-results!! ctx retriggered-build-number original-build-result)
-      (assoc original-step-result
-        :retrigger-mock-for-build-number retriggered-build-number))))
-
-(defn- clear-retrigger-data [ctx]
-  (assoc ctx
-    :retriggered-build-number nil
-    :retriggered-step-id nil))
-
-(defn- replace-step-with-retrigger-mock [retrigger-predicate [ctx step]]
-  (let [retriggered-build-number (:retriggered-build-number ctx)]
-    (case (retrigger-predicate ctx step)
-      :rerun [ctx step]
-      :run [(clear-retrigger-data ctx) step]
-      :mock [ctx (retrigger-mock-step retriggered-build-number)])))
-
-(defn- add-retrigger-mocks [retrigger-predicate root-ctx step-contexts]
-  (if (:retriggered-build-number root-ctx)
-    (map (partial replace-step-with-retrigger-mock retrigger-predicate) step-contexts)
-    step-contexts))
-
 (defn- call-step-result-producer [step-result-producer]
   (fn [step-contexts-and-steps args _]
     (let [step-results (step-result-producer args step-contexts-and-steps)]
@@ -111,22 +70,6 @@
       result)))
 
 
-(defn- add-kill-switch [is-killed]
-  (fn [[ctx step]]
-    [(assoc ctx :is-killed is-killed) step]))
-
-(defn- wrap-kill-handling [handler is-killed]
-  (fn [step-contexts-and-steps args ctx]
-    (handler (map (add-kill-switch is-killed) step-contexts-and-steps)
-             args
-             ctx)))
-
-(defn- wrap-retrigger-handling [handler retrigger-predicate]
-  (fn [step-contexts-and-steps args ctx]
-    (handler (add-retrigger-mocks retrigger-predicate ctx step-contexts-and-steps)
-             args
-             ctx)))
-
 (defn- wrap-filter-nil-steps [handler]
   (fn [step-contexts-and-steps args ctx]
     (handler (filter (fn [[_ step]] (not-nil? step)) step-contexts-and-steps)
@@ -138,10 +81,10 @@
                                        :or   {step-result-producer (serial-step-result-producer)
                                               is-killed            (atom false)
                                               unify-results-fn     (unify-only-status status/successful-when-all-successful)
-                                              retrigger-predicate  sequential-retrigger-predicate}}]
+                                              retrigger-predicate  retrigger/sequential-retrigger-predicate}}]
   (let [handler-chain (-> (call-step-result-producer step-result-producer)
                           (wrap-inheritance unify-results-fn)
-                          (wrap-kill-handling is-killed)
-                          (wrap-retrigger-handling retrigger-predicate)
+                          (kill/wrap-execute-steps-with-kill-handling is-killed)
+                          (retrigger/wrap-retrigger-handling retrigger-predicate)
                           (wrap-filter-nil-steps))]
     (handler-chain (contexts-for-steps steps ctx) args ctx)))